APIs/gitlab.com/v3/python-aiohttp-client/tests/test_templates_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.repo_license import RepoLicense
from openapi_server.models.template import Template
from openapi_server.models.templates_list import TemplatesList


pytestmark = pytest.mark.asyncio

async def test_get_v3_templates_dockerfiles(client):
    """Test case for get_v3_templates_dockerfiles

    Get the list of the available template
    """
    headers = { 
        'Accept': 'application/json',
        'private_token_query': 'special-key',
        'private_token_header': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/v3/templates/dockerfiles',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_v3_templates_dockerfiles_name(client):
    """Test case for get_v3_templates_dockerfiles_name

    Get the text for a specific template present in local filesystem
    """
    headers = { 
        'Accept': 'application/json',
        'private_token_query': 'special-key',
        'private_token_header': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/v3/templates/dockerfiles/{name}'.format(name='name_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_v3_templates_gitignores(client):
    """Test case for get_v3_templates_gitignores

    Get the list of the available template
    """
    headers = { 
        'Accept': 'application/json',
        'private_token_query': 'special-key',
        'private_token_header': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/v3/templates/gitignores',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_v3_templates_gitignores_name(client):
    """Test case for get_v3_templates_gitignores_name

    Get the text for a specific template present in local filesystem
    """
    headers = { 
        'Accept': 'application/json',
        'private_token_query': 'special-key',
        'private_token_header': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/v3/templates/gitignores/{name}'.format(name='name_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_v3_templates_gitlab_ci_ymls(client):
    """Test case for get_v3_templates_gitlab_ci_ymls

    Get the list of the available template
    """
    headers = { 
        'Accept': 'application/json',
        'private_token_query': 'special-key',
        'private_token_header': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/v3/templates/gitlab_ci_ymls',
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_v3_templates_gitlab_ci_ymls_name(client):
    """Test case for get_v3_templates_gitlab_ci_ymls_name

    Get the text for a specific template present in local filesystem
    """
    headers = { 
        'Accept': 'application/json',
        'private_token_query': 'special-key',
        'private_token_header': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/v3/templates/gitlab_ci_ymls/{name}'.format(name='name_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_v3_templates_licenses(client):
    """Test case for get_v3_templates_licenses

    Get the list of the available license template
    """
    params = [('popular', True)]
    headers = { 
        'Accept': 'application/json',
        'private_token_query': 'special-key',
        'private_token_header': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/v3/templates/licenses',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_v3_templates_licenses_name(client):
    """Test case for get_v3_templates_licenses_name

    Get the text for a specific license
    """
    headers = { 
        'Accept': 'application/json',
        'private_token_query': 'special-key',
        'private_token_header': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/v3/templates/licenses/{name}'.format(name='name_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

